requirements.txt
requests
beautifulsoup4
google-generativeai
feedparser
facebook-sdk